requirements.txt
compel==2.0.3
# typing-extensions==4.9.0
typing_extensions==4.11.0
pydantic==2.5.3
modelscope==1.14.0
numpy==1.26.3
diffusers==0.32.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.2.0
peft==0.11.1
wandb==0.21.0
protobuf==5.28.3
Requests==2.31.0
tensorflow==2.18.0
transformers==4.50.0
transformers-stream-generator==0.0.5
# torchao==0.8.0
addict==2.4.0
datasets==2.18.0
sentencepiece==0.2.0
simplejson==3.19.3
sortedcontainers==2.4.0
accelerate==0.29.3
kornia==0.7.4
ftfy==6.3.1
regex
tqdm
bert_score
decord
liger_kernel
clint
clip
timm
fairscale
# pdb
# random
# xformers==0.0.23